requirements.txt
pdfplumber>=0.11,<1
pypdf>=5,<6
pytest>=8,<9
fastapi>=0.115,<1
uvicorn>=0.30,<1
openai>=1.54,<2
python-dotenv>=1.0,<2
